move actions-column from grid to file-manager. small adjustments.
