+++ src/components/FileManager/constants.tsx
@@ -33,8 +33,6 @@ export const FOLDERS_TREE_PANEL_MAX_WIDTH = 460;
 
 export const COMPACT_VIEW_HEADER_HEIGHT = 44;
 
-export const COMPACT_VIEW_FOLDER_ROW_HEIGHT = COMPACT_VIEW_HEADER_HEIGHT;
-
 export const COMPACT_VIEW_FILE_ROW_HEIGHT = 56;
 
 export const DEFAULT_COMPACT_VIEW_WIDTH_BREAKPOINT = 800;
